+++ engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/source/BaseKafkaAvroSourceFactory.scala
@@ -1,8 +1,8 @@
 package pl.touk.nussknacker.engine.avro.source
 
+import org.apache.flink.streaming.connectors.kafka.KafkaDeserializationSchema
 import pl.touk.nussknacker.engine.api.MetaData
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.NodeId
-import pl.touk.nussknacker.engine.api.test.TestParsingUtils
 import pl.touk.nussknacker.engine.api.typed.{ReturningType, typing}
 import pl.touk.nussknacker.engine.avro.serialization.KafkaAvroDeserializationSchemaFactory
 import pl.touk.nussknacker.engine.avro.typed.AvroSchemaTypeDefinitionExtractor
@@ -24,29 +24,39 @@ abstract class BaseKafkaAvroSourceFactory[T: ClassTag](timestampAssigner: Option
                    kafkaConfig: KafkaConfig,
                    deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory,
                    createRecordFormatter: RecordFormatter,
-                   schemaDeterminer: AvroSchemaDeterminer,
+                   keySchemaDeterminer: AvroSchemaDeterminer,
+                   valueSchemaDeterminer: AvroSchemaDeterminer,
                    returnGenericAvroType: Boolean)
                   (implicit processMetaData: MetaData,
                    nodeId: NodeId): KafkaSource[T] = {
 
-    val schemaData = schemaDeterminer.determineSchemaUsedInTyping.valueOr(SchemaDeterminerErrorHandler.handleSchemaRegistryErrorAndThrowException)
-    val schemaUsedInRuntime = schemaDeterminer.toRuntimeSchema(schemaData)
+    // key schema
+    val keySchemaData = keySchemaDeterminer.determineSchemaUsedInTyping.valueOr(SchemaDeterminerErrorHandler.handleSchemaRegistryErrorAndThrowException)
+    val keySchemaDataUsedInRuntime = keySchemaDeterminer.toRuntimeSchema(keySchemaData)
+
+    // value schema
+    val valueSchemaData = valueSchemaDeterminer.determineSchemaUsedInTyping.valueOr(SchemaDeterminerErrorHandler.handleSchemaRegistryErrorAndThrowException)
+    val valueSchemaUsedInRuntime = valueSchemaDeterminer.toRuntimeSchema(valueSchemaData)
+
+    // prepare KafkaDeserializationSchema based on key and value schema
+    // TODO: add key-value deserialization as default scenario: create[K, V]
+    val deserializationSchema = deserializationSchemaFactory.create[Any, T](kafkaConfig, keySchemaDataUsedInRuntime, valueSchemaUsedInRuntime).asInstanceOf[KafkaDeserializationSchema[T]]
 
     if (returnGenericAvroType) {
       new KafkaSource(
         List(preparedTopic),
         kafkaConfig,
-        deserializationSchemaFactory.create[T](schemaUsedInRuntime, kafkaConfig),
+        deserializationSchema,
         assignerToUse(kafkaConfig),
         createRecordFormatter
       ) with ReturningType {
-        override def returnType: typing.TypingResult = AvroSchemaTypeDefinitionExtractor.typeDefinition(schemaData.schema)
+        override def returnType: typing.TypingResult = AvroSchemaTypeDefinitionExtractor.typeDefinition(valueSchemaData.schema)
       }
     } else {
       new KafkaSource(
         List(preparedTopic),
         kafkaConfig,
-        deserializationSchemaFactory.create[T](schemaUsedInRuntime, kafkaConfig),
+        deserializationSchema,
         assignerToUse(kafkaConfig),
         createRecordFormatter
       )